# docs/swagger.yaml
basePath: /v0
definitions:
  inputs.AlertBody:
    properties:
      message:
        type: string
      parent_id:
        type: integer
      severity:
        type: integer
    required:
    - message
    type: object
...
  inputs.AlertUpdate:
    properties:
      message:
        type: string
      severity:
        type: integer
    type: object
...
  models.Response:
    properties:
      data:
        type: object
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: I have no specific description
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hod Hod API
  version: "1.0"
paths:
  /alert:
    get:
      description: return alerts
      operationId: get_alerts
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get alerts
      tags:
      - Alert
    post:
      description: create alert
      operationId: new_alert
      parameters:
      - description: alert
        in: body
        name: alertBody
        required: true
        schema:
          $ref: '#/definitions/inputs.AlertBody'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: new alert
      tags:
      - Alert
...
  /alert/{id}:
    delete:
      description: delete alert
      operationId: delete_alert
      parameters:
      - description: alert_id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: delete alert
      tags:
      - Alert
    get:
      description: return alert
      operationId: get_alert
      parameters:
      - description: alert_id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: get alert
      tags:
      - Alert
    put:
      description: update alert
      operationId: update_alert
      parameters:
      - description: alert_id
        in: path
        name: id
        required: true
        type: integer
      - description: alert
        in: body
        name: alertBody
        required: true
        schema:
          $ref: '#/definitions/inputs.AlertUpdate'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            type: json
      security:
      - ApiKeyAuth: []
      summary: update alert
      tags:
      - Alert
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"